smart_bill_bot/requirements.txt
flask==2.3.2
transformers==4.31.0
datasets==2.14.5
torch==2.0.0
pytesseract==0.3.10
pdf2image==1.16.3
pillow==9.5.0
numpy==1.23.5
